requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dateutil==2.9.0.post0
pytz==2024.2
jinja2==3.1.4
resend==1.0.0
